// FE/src/app/shared/models/DeviceManager/device-parameter-use.model.ts
export class DeviceParameterUse {
    id?: number;
    value?: number;
    min?: number;
    max?: number;
    unit?: string;
    description?: string;
    createdAt?: Date;
    updatedAt?: Date;
    createdBy?: string;
    updatedBy?: string;
    status?: number; 
    device?: any
    parameter?: any
}